Added improved arduino communication to the Android app.
