+++ app/src/main/res/layout/administrator_activity.xml
@@ -1,105 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
-
-<TableLayout
+<RelativeLayout
 
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="horizontal"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-    android:layout_height="0dp"
-    android:weightSum="2">
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:weightSum="1"
+    android:keepScreenOn="true">
 
-    <TableRow>
-        <ListView
-            xmlns:android="http://schemas.android.com/apk/res/android"
-            android:id="@+id/purchase_listview"
-            android:layout_width="600dp"
-            android:layout_weight="1"
-            android:layout_height="500dp"
-            android:layout_marginLeft="20dp"
-            android:layout_gravity="bottom|left|top" />
+    <android.support.v4.view.ViewPager
+        android:id="@+id/container"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
 
-            <LinearLayout
-                android:layout_weight="1"  >
 
-                <TextView
-                    android:id="@+id/date_view"
-                    android:layout_width="250dp"
-                    android:layout_height="wrap_content"
-                    android:text="@string/admin_date_view"
-                    android:gravity="center"
-                    android:textStyle="bold"
-                    android:textSize="20sp"/>
-                <TextView
-                    android:id="@+id/start_date"
-                    android:layout_width="100dp"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="80dp"
-                    android:layout_marginLeft="-300dp"
-                    android:text="@string/start_Date"
-                    android:gravity="center"
-                    android:textStyle="bold"
-                    android:textSize="20sp"/>
-                <EditText
-                    android:id="@+id/start_tag"
-                    android:ems="10"
-                    android:layout_marginLeft="-50dp"
-                    android:layout_marginTop="120dp"
-                    android:layout_width="50dp"
-                    android:layout_height="wrap_content"
-                    android:inputType="text"
-                    android:layout_weight="1" />
-                <Button
-                    android:id="@+id/filter_list_tag"
-                    android:layout_width="50dp"
-                    android:layout_height="50dp"
-                    android:layout_gravity="center"
-                    android:background="@drawable/buttonshape"
-                    android:text="@string/filter"
-                    android:clickable="true"/>
-                <TextView
-                    android:id="@+id/end_date"
-                    android:layout_width="100dp"
-                    android:layout_height="wrap_content"
-                    android:text="@string/end_Date"
-                    android:gravity="center"
-                    android:layout_marginTop="80dp"
-                    android:layout_marginLeft="20dp"
-                    android:textStyle="bold"
-                    android:textSize="20sp"/>
-                <EditText
-                    android:id="@+id/end_tag"
-                    android:layout_width="50dp"
-                    android:layout_height="wrap_content"
-                    android:inputType="text"
-                    android:ems="10"
-                    android:layout_marginLeft="-100dp"
-                    android:layout_marginTop="120dp"
-                    android:layout_weight="1" />
-                <Button
-                    android:layout_width="120dp"
-                    android:layout_height="120dp"
-                    android:id="@+id/quit_button"
-                    android:layout_gravity="bottom"
-                    android:layout_marginLeft="-300dp"
-                    android:background="@drawable/buttonshape"
-                    android:text="Quit" />
-                <Button
-                    android:layout_width="120dp"
-                    android:layout_height="120dp"
-                    android:id="@+id/export_button"
-                    android:layout_marginTop="350dp"
-                    android:layout_marginLeft="10dp"
-                    android:background="@drawable/buttonshape"
-                    android:text="Export" />
-                <Button
-                    android:layout_width="120dp"
-                    android:layout_height="120dp"
-                    android:id="@+id/close_button"
-                    android:layout_marginTop="350dp"
-                    android:layout_marginLeft="10dp"
-                    android:background="@drawable/buttonshape"
-                    android:text="Close" />
-            </LinearLayout>
-    </TableRow>
-</TableLayout>
+</RelativeLayout>
