+++ app/src/main/res/layout/administrator_log_export_view.xml
@@ -5,8 +5,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="horizontal"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
-    android:layout_height="0dp"
+    android:layout_height="match_parent"
     android:weightSum="2">
 
     <TableRow>
